+++ src/requires/controlclass.lua
@@ -761,8 +761,12 @@ function ControlClass(Nav, c, u, s, atlas, vBooster, hover, antigrav, shield_1, 
         msgText = "Extra Engine Tags: "..UseExtra 
         u.stopTimer("tagTick")
     end
-    -- UNCOMMENT BELOW LINE TO ACTIVATE A CUSTOM OVERRIDE FILE TO OVERRIDE SPECIFIC FUNCTIONS
-    --for k,v in pairs(require("autoconf/custom/archhud/custom/customcontrolclass")) do Control[k] = v end 
+
+    -- TO ACTIVATE A CUSTOM OVERRIDE FILE TO OVERRIDE SPECIFIC FUNCTIONS THE FOLLOWING FILE MUST BE PRESENT
+    local _, ccOverride = pcall(require, "autoconf/custom/archhud/custom/customcontrolclass")
+    if type(ccOverride) == "table" then
+        for k,v in pairs(ccOverride) do Control[k] = v end 
+    end    
 
     return Control
 end